move and rename files and add macro parameters
