+++ launch/sonar_single_beam_basic.launch
@@ -2,12 +2,12 @@
 <launch>
   <!-- world -->
   <include file="$(find gazebo_ros)/launch/empty_world.launch">
-    <arg name="world_name" value="$(find nps_uw_sensors)/worlds/sonar_single_beam_basic.world"/>
+    <arg name="world_name" value="$(find nps_uw_sensors_gazebo)/worlds/sonar_single_beam_basic.world"/>
   </include>
 
 <!-- sonar_single_beam_robot xacro -->
   <param name="sonar_single_beam_robot"
-         command="$(find xacro)/xacro '$(find nps_uw_sensors)/urdf/sonar_single_beam_robot.xacro'"/>
+         command="$(find xacro)/xacro '$(find nps_uw_sensors_gazebo)/urdf/sonar_single_beam_robot.xacro'"/>
 
 <!-- spawn sonar_single_beam_robot -->
   <node name="urdf_spawner" pkg="gazebo_ros" type="spawn_model"
