+++ models/blueview_p900/model.config
@@ -1,0 +1,15 @@
+<?xml version="1.0"?>
+<model>
+  <name>blueview_p900</name>
+  <version>1.0</version>
+  <sdf version="1.6">model.sdf</sdf>
+
+  <author>
+    <name>Brian Bingham</name>
+    <email>briansbingham@gmail.com</email>
+  </author>
+
+  <description>
+    Blueview P900
+  </description>
+</model>
