+++ models/blueview_p900/model.sdf
@@ -1,0 +1,76 @@
+<?xml version="1.0" ?>
+<sdf version="1.6">
+  <model name="blueview_p900">
+    <link name="forward_sonar_link">
+      <inertial>
+        <pose>0 0 0 0 0 0</pose>
+        <mass>1</mass>
+        <!-- Mark only - based on cylinder -->
+        <inertia>
+          <ixx>0.1</ixx>
+          <ixy>0</ixy>
+          <ixz>0</ixz>
+          <iyy>0.1</iyy>
+          <iyz>0</iyz>
+          <izz>0.1</izz>
+        </inertia>
+      </inertial>
+      <collision name="collision">
+        <pose>-0.08 0 0 0 1.571 0</pose>
+        <geometry>
+          <cylinder>
+            <radius>0.07</radius>
+            <length>0.28</length>
+          </cylinder>
+        </geometry>
+      </collision>
+      <visual name="visual">
+        <geometry>
+	  <mesh>
+	    <!--<uri>file://$(find uuv_sensor_ros_plugins)/meshes/p900.dae</uri>-->
+	    <uri>model://blueview_p900/meshes/p900.dae</uri>
+	  </mesh>
+        </geometry>
+      </visual>
+
+      <sensor name="image_sonar" type="depth">
+	<visualize>true</visualize>
+	<always_on>true</always_on>
+	<update_rate>40</update_rate>
+	<camera>
+	  <!-- 90 degrees for the M900-90 -->
+	  <horizontal_fov>1.54719755</horizontal_fov>
+          <image>
+	    <width>760</width>
+	    <!-- Set vertical FOV by setting image height -->
+	    <!-- For 20 deg. -->
+	    <height>169</height>
+            <format>R8G8B8</format>
+          </image>
+          <clip>
+            <near>0.1</near>
+            <far>17</far>
+          </clip>
+	  <depth_camera>
+            <output>depths</output>
+          </depth_camera>
+          <save enabled="true">
+            <path>/tmp/camera</path>
+          </save>
+	</camera>
+	<plugin filename="libimage_sonar_ros_plugin.so" name="forward_sonar">
+	  <!-- This name is prepended to ROS topics -->
+	  <cameraName>blueview_p900</cameraName>
+	  <!-- ROS publication topics --> 
+	  <imageTopicName>image_raw</imageTopicName>
+	  <cameraInfoTopicName>image_raw/camera_info</cameraInfoTopicName>
+	  <pointCloudTopicName>point_cloud</pointCloudTopicName>
+	  <depthImageTopicName>image_depth</depthImageTopicName>
+	  <depthImageCameraInfoTopicName>image_depth/camera_info</depthImageCameraInfoTopicName>
+
+          <frameName>forward_sonar_optical_link</frameName>
+	</plugin>
+      </sensor>     
+    </link>
+  </model>
+</sdf>
