+++ package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>nps_uw_sensors</name>
+  <name>nps_uw_sensors_gazebo</name>
   <version>0.0.1</version>
   <description>
     NPS Gazebo underwater sensor plugins.
@@ -12,32 +12,11 @@
 
   <buildtool_depend>catkin</buildtool_depend>
 
-  <build_depend>gazebo_dev</build_depend>
-  <exec_depend>gazebo_dev</exec_depend>
-
   <depend>gazebo_msgs</depend>
-  <depend>geometry_msgs</depend>
   <depend>sensor_msgs</depend>
-  <depend>trajectory_msgs</depend>
-  <depend>std_srvs</depend>
+  <depend>std_msgs</depend>
   <depend>roscpp</depend>
   <depend>rospy</depend>
-  <depend>nodelet</depend>
-  <depend>angles</depend>
-  <depend>nav_msgs</depend>
-  <depend>urdf</depend>
   <depend>tf</depend>
-  <depend>tf2_ros</depend>
-  <depend>dynamic_reconfigure</depend>
-  <depend>rosgraph_msgs</depend>
-  <depend>image_transport</depend>
-  <depend>rosconsole</depend>
-  <build_depend>message_generation</build_depend>
-  <exec_depend>message_runtime</exec_depend>
-  <depend>cv_bridge</depend>
-  <depend>polled_camera</depend>
-  <depend>diagnostic_updater</depend>
-  <depend>camera_info_manager</depend>
-  <depend>std_msgs</depend>
 
 </package>
