+++ urdf/sonar_single_beam.xacro
@@ -1,6 +1,8 @@
 <?xml version="1.0"?>
 <robot xmlns:xacro="http://www.ros.org/wiki/xacro">
-  <xacro:macro name="sonar_single_beam" params="namespace suffix parent_link topic *origin">
+
+  <!-- note: a samples value less than 3 causes Ogre::AxisAlignedBox error -->
+  <xacro:macro name="sonar_single_beam" params="namespace suffix parent_link topic beam_width samples update_rate min_range max_range *origin">
 
     <!-- joint to Sonar link -->
     <joint name="${namespace}_sonar_single_beam${suffix}_joint" type="fixed">
@@ -24,7 +26,7 @@
       </collision>
       <visual name="visual">
         <geometry>
-          <mesh filename="file://$(find nps_uw_sensors)/meshes/blueview_p900/p900.dae"/>
+          <mesh filename="file://$(find nps_uw_sensors_gazebo)/models/blueview_p900/meshes/p900.dae"/>
         </geometry>
       </visual>
     </link>
@@ -35,16 +37,23 @@
  
     <sensor type="gpu_ray" name="ray_sensor">
       <always_on>true</always_on>
-      <update_rate>15.0</update_rate>
+      <update_rate>${update_rate}</update_rate>
       <pose>0 0 0 0 0 0</pose>
       <visualize>true</visualize>
       <ray>
         <scan>
           <horizontal>
-            <samples>3</samples>
+            <samples>${samples}</samples>
+            <!-- samples value less than 3 causes Ogre::AxisAlignedBox error -->
+            <xacro:if value="${samples == '1'}">
+              <min_angle>0.0</min_angle>
+              <max_angle>0.0</max_angle>                        
+            </xacro:if>
+            <xacro:unless value="${samples == '1'}">
+              <min_angle>${-beam_width / 2.0}</min_angle>
+              <max_angle>${beam_width / 2.0}</max_angle>                        
+            </xacro:unless>
             <resolution>1</resolution>            
-            <min_angle>-0.1</min_angle>
-            <max_angle>0.1</max_angle>                        
           </horizontal>        
           <vertical>            
             <samples>1</samples>
@@ -55,8 +64,8 @@
         </scan>
 
         <range>
-          <min>0.1</min>
-          <max>10.0</max>
+          <min>${min_range}</min>
+          <max>${max_range}</max>
         </range>
       </ray>
