+++ urdf/sonar_single_beam_robot.xacro
@@ -2,7 +2,7 @@
 <robot name="sonar_single_beam_robot"
        xmlns:xacro="http://www.ros.org/wiki/xacro">
 
-  <xacro:include filename="$(find nps_uw_sensors)/urdf/sonar_single_beam.xacro" />
+  <xacro:include filename="$(find nps_uw_sensors_gazebo)/urdf/sonar_single_beam.xacro" />
 
   <link name="world"/>
 
@@ -10,7 +10,12 @@
             namespace="sonar_single_beam_robot"
             suffix="1"
             parent_link="world"
-            topic="sonar_single_beam">
+            topic="sonar_single_beam"
+            beam_width="0.2"
+            samples="3"
+            update_rate="15"
+            min_range="0.1"
+            max_range="7.0">
     <origin xyz="2.1 1.0 1.0" rpy="0 0 1.5709"/>
   </xacro:sonar_single_beam>
